requirements.txt
# Example: Add your project dependencies here
GitPython==3.1.8
pydantic-ai-slim[openai]==0.3.1
# flask==2.0.1
